+++ tests/src/test/scala/whisk/core/cli/test/SwiftTests.scala
@@ -90,4 +90,84 @@ class SwiftTests
                     result.toString should not include ("Error")
             }
     }
+
+    behavior of "Swift 3 Whisk SDK tests"
+
+    it should "allow Swift actions to invoke other actions" in withAssetCleaner(wskprops) {
+        (wp, assetHelper) =>
+            // use CLI to create action from dat/actions/invokeAction.swift
+            val file = TestUtils.getTestActionFilename("invoke.swift")
+            val actionName = "invokeAction"
+            assetHelper.withCleaner(wsk.action, actionName) {
+                (action, _) => action.create(name = actionName, artifact = Some(file), kind = Some("swift:3"))
+            }
+
+            // invoke the action
+            val run = wsk.action.invoke(actionName)
+            withActivation(wsk.activation, run, initialWait = 5 seconds, totalWait = 60 seconds) {
+                activation =>
+                    // should be successful
+                    activation.response.success shouldBe true
+
+                    // should have a field named "activationId" which is the date action's activationId
+                    activation.response.result.get.fields("activationId").toString.length should be >= 32
+
+                // check for "date" field that comes from invoking the date action
+                //activation.response.result.get.fieldPathExists("response", "result", "date") should be(true)
+            }
+    }
+
+    it should "allow Swift actions to invoke other actions and not block" in withAssetCleaner(wskprops) {
+        (wp, assetHelper) =>
+            // use CLI to create action from dat/actions/invokeNonBlocking.swift
+            val file = TestUtils.getTestActionFilename("invokeNonBlocking.swift")
+            val actionName = "invokeNonBlockingAction"
+            assetHelper.withCleaner(wsk.action, actionName) {
+                (action, _) => action.create(name = actionName, artifact = Some(file), kind = Some("swift:3"))
+            }
+
+            // invoke the action
+            val run = wsk.action.invoke(actionName)
+            withActivation(wsk.activation, run, initialWait = 5 seconds, totalWait = 60 seconds) {
+                activation =>
+                    // should not have a "response"
+                    whisk.utils.JsHelpers.fieldPathExists(activation.response.result.get, "response") shouldBe false
+
+                    // should have a field named "activationId" which is the date action's activationId
+                    activation.response.result.get.fields("activationId").toString.length should be >= 32
+            }
+    }
+
+    it should "allow Swift actions to trigger events" in withAssetCleaner(wskprops) {
+        (wp, assetHelper) =>
+            // create a trigger
+            val triggerName = s"TestTrigger ${System.currentTimeMillis()}"
+            assetHelper.withCleaner(wsk.trigger, triggerName) {
+                (trigger, _) => trigger.create(triggerName)
+            }
+
+            // create an action that fires the trigger
+            val file = TestUtils.getTestActionFilename("trigger.swift")
+            val actionName = "ActionThatTriggers"
+            assetHelper.withCleaner(wsk.action, actionName) {
+                (action, _) => action.create(name = actionName, artifact = Some(file), kind = Some("swift:3"))
+            }
+
+            // invoke the action
+            val run = wsk.action.invoke(actionName, Map("triggerName" -> triggerName.toJson))
+            withActivation(wsk.activation, run, initialWait = 5 seconds, totalWait = 60 seconds) {
+                activation =>
+                    // should be successful
+                    activation.response.success shouldBe true
+
+                    // should have a field named "activationId" which is the date action's activationId
+                    activation.response.result.get.fields("activationId").toString.length should be >= 32
+
+                    // should result in an activation for triggerName
+                    val triggerActivations = wsk.activation.pollFor(1, Some(triggerName), retries = 20)
+                    withClue(s"trigger activations for $triggerName:") {
+                        triggerActivations.length should be(1)
+                    }
+            }
+    }
 }
